Add sample rate option to config, and better output file name
The native sample rate of the synthesized voice audio can now be set in the config file. This is useful for services like Azure's TTS which has voices that can be output in 48KHz.

Also now creates the final output file name based on the name of the input video file.