Remove need for temporary files
The program now processes the audio files (such as trimming and stretching) all in memory instead of having to read and write files. Though it will still download and save the TTS files sent by the APIs though (In the workingFolder)

Also removed requirement for librosa, because now use pydub to get the audio duration.