Implement DeepL translation
Implement DeepL translation service, including its logic for batch translation as well as its settings. This introduces the python package deepl which is the official package for the API. Documented DeepL in the README, add logic to not require deepl key if using google translate, and viceversa, not requiring google api key or credentials if using deepl + azure.