Implement Azure Exact Duration Feature
Azure TTS now supports setting the exact desired duration of the synthesized audio clip. This eliminates the need for doing two passes or stretching the audio.

Also set tailing and leading silence to 0ms so trimming wouldn't be necessary, but these tags to not appear to fully work at the moment.