Fix Fragmented Translations - Google
Had to create quite a significant workaround to prevent it from fragmenting sentences while also being able to mark the splits in lines. I add <b> tags to the combined string to translate, which Google Translate maintains. But it seems to have a limit of 50 of these, so I had to make it so it splits chunks at 50 lines.