Add auto-calculated char rate goal with config setting
The script will now auto calculate the average character-rate of the entire video and use that as the char rate goal, instead of hard coding it at 20 like before. The user can change this behavior in the config file with the "speech_rate_goal" setting.

Add message when downloading

Also made it so now when in debug mode, it will put out a proper completed subtitle file, and a separate debug version of the file.